+++ .pipelines/ci/templates/build-powertoys-ci.yml
