+++ .pipelines/ci/templates/build-powertoys-steps.yml
@@ -73,29 +73,34 @@ steps:
     maximumCpuCount: true
 
 
-# directly not doing WinAppDriver testing
+# Running VSTest@2 step seperatly for .netframework and .netcore.
+# .NetCore assemblies
 - task: VSTest@2
+  displayName: 'Run .Net Core Tests'
   inputs:
     platform: '$(BuildPlatform)'
     configuration: '$(BuildConfiguration)'
     testSelector: 'testAssemblies'
     testAssemblyVer2: |
-      !**\*PowerToysTests.dll
-      !**\*TestAdapter.dll
+      #**\Microsoft.Plugin.Program.UnitTests.dll
+      #**\Microsoft.Plugin.Uri.UnitTests.dll
+      #**\Wox.Test.dll
+      **\*Microsoft.PowerToys.Settings.UI.UnitTests.dll
       !**\obj\**
+# .NetFramework assemblies
+- task: VSTest@2
+  displayName: 'Run .Net Framework Tests'
+  inputs:
+    platform: '$(BuildPlatform)'
+    configuration: '$(BuildConfiguration)'
+    testSelector: 'testAssemblies'
+    testAssemblyVer2: |
       **\ImageResizer.Test.dll
       **\KeyboardManagerTest.dll
-      **\Microsoft.Plugin.Program.UnitTests.dll
-      **\Microsoft.Plugin.Uri.UnitTests.dll
       **\PowerRenameUnitTests.dll
       **\UnitTests-CommonLib.dll
       **\PreviewPaneUnitTests.dll #this is the markdown tests
       **\UnitTests-PreviewHandlerCommon.dll
       **\UnitTests-SvgPreviewHandler.dll
-      **\Wox.Test.dll
       **\powerpreviewTest.dll
-      # Can't get working, says no tests
-      #   #**\Microsoft.PowerToys.Settings.UnitTest.exe
-      # crashes CI
-      #**\interop-tests.dll 
-      #**\UnitTests-FancyZones.dll
+      !**\obj\**
