+++ src/core/Microsoft.PowerToys.Settings.UI.UnitTests/General.cs
@@ -1,0 +1,32 @@
+﻿// Copyright (c) Microsoft Corporation
+// The Microsoft Corporation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.IO;
+using System.Text.Json;
+using Microsoft.PowerToys.Settings.UI.Lib;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace ViewModelTests
+{
+    [TestClass]
+    public class SampleTest
+    {
+        [TestInitialize]
+        public void Setup()
+        {
+        }
+
+        [TestCleanup]
+        public void CleanUp()
+        {
+        }
+
+        [TestMethod]
+        public void SampleTest_For_Settings()
+        {
+
+        }
+    }
+}
