--- src/core/Microsoft.PowerToys.Settings.UnitTest/ModelsTests/BasePTSettingsTest.cs
@@ -1,17 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation
-// The Microsoft Corporation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using Microsoft.PowerToys.Settings.UI.Lib;
-
-namespace Microsoft.PowerToys.Settings.UnitTest
-{
-    public class BasePTSettingsTest : BasePTModuleSettings
-    {
-        public BasePTSettingsTest()
-        {
-            Name = string.Empty;
-            Version = string.Empty;
-        }
-    }
-}
