--- src/core/Microsoft.PowerToys.Settings.UnitTest/UnitTestApp.xaml
@@ -1,7 +1,0 @@
-﻿<Application
-    x:Class="Microsoft.PowerToys.Settings.UnitTest.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Microsoft.PowerToys.Settings.UnitTest">
-
-</Application>
